refactor: FilterCategoryList 컴포넌트 분리
